extract workflow action routes
