+++ src/api/routers/rebalance_runs_workflow_action_routes.py
@@ -1,0 +1,240 @@
+from typing import Annotated, Any, Optional
+
+from fastapi import Header, HTTPException, Path, Request, status
+
+from src.api.observability import record_workflow_decision
+from src.api.routers import rebalance_runs as shared
+from src.core.rebalance_runs import (
+    DpmRunNotFoundError,
+    DpmRunSupportService,
+    DpmRunWorkflowActionRequest,
+    DpmRunWorkflowResponse,
+    DpmWorkflowDisabledError,
+    DpmWorkflowTransitionError,
+)
+from src.core.rebalance_runs.models import DpmWorkflowActionType
+
+
+_WORKFLOW_ACTION_DESCRIPTION = (
+    "Applies one workflow action (`APPROVE`, `REJECT`, `REQUEST_CHANGES`) for a discretionary "
+    "mandate rebalance run and returns updated workflow state. Supply the reviewer action in the "
+    "request body and optional `X-Correlation-Id` header for action traceability. This endpoint "
+    "does not accept query parameters."
+)
+
+_RouteResponses = dict[int | str, dict[str, Any]]
+
+_WORKFLOW_ACTION_RESPONSES: _RouteResponses = {
+    200: {"description": "Updated workflow state after applying the reviewer action."},
+    404: {"description": "Workflow disabled, run not found, or idempotency mapping not found."},
+    409: {"description": "Workflow action is not valid for the current run state."},
+    422: {"description": "Invalid action payload or unsupported query parameters were supplied."},
+}
+
+
+def _record_workflow_action_metric(
+    *,
+    surface: str,
+    action: DpmWorkflowActionType,
+    outcome: str,
+) -> None:
+    record_workflow_decision(
+        surface=surface,
+        action=action.lower(),
+        outcome=outcome,
+    )
+
+
+@shared.router.post(
+    "/rebalance/runs/{rebalance_run_id}/workflow/actions",
+    response_model=DpmRunWorkflowResponse,
+    status_code=status.HTTP_200_OK,
+    summary="Apply lotus-manage Run Workflow Action",
+    description=_WORKFLOW_ACTION_DESCRIPTION,
+    responses=_WORKFLOW_ACTION_RESPONSES,
+)
+def apply_dpm_run_workflow_action(
+    request: Request,
+    rebalance_run_id: Annotated[
+        str,
+        Path(description="lotus-manage run identifier.", examples=["rr_abc12345"]),
+    ],
+    payload: DpmRunWorkflowActionRequest,
+    service: DpmRunSupportService = shared.Depends(shared.get_dpm_run_support_service),
+    x_correlation_id: Annotated[
+        Optional[str],
+        Header(
+            description="Optional correlation id for workflow action request tracing.",
+            examples=["corr-workflow-001"],
+        ),
+    ] = None,
+) -> DpmRunWorkflowResponse:
+    shared._assert_support_apis_enabled()
+    shared._assert_workflow_enabled()
+    shared._reject_unexpected_query_params(request, allowed_params=set())
+    try:
+        response = service.apply_workflow_action(
+            rebalance_run_id=rebalance_run_id,
+            action=payload.action,
+            reason_code=payload.reason_code,
+            comment=payload.comment,
+            actor_id=payload.actor_id,
+            correlation_id=x_correlation_id or "c_none",
+        )
+        _record_workflow_action_metric(surface="run", action=payload.action, outcome="success")
+        return response
+    except DpmRunNotFoundError as exc:
+        _record_workflow_action_metric(surface="run", action=payload.action, outcome="not_found")
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=str(exc)) from exc
+    except DpmWorkflowDisabledError as exc:
+        _record_workflow_action_metric(surface="run", action=payload.action, outcome="disabled")
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=str(exc)) from exc
+    except DpmWorkflowTransitionError as exc:
+        _record_workflow_action_metric(surface="run", action=payload.action, outcome="conflict")
+        raise HTTPException(status_code=status.HTTP_409_CONFLICT, detail=str(exc)) from exc
+
+
+@shared.router.post(
+    "/rebalance/runs/by-correlation/{correlation_id}/workflow/actions",
+    response_model=DpmRunWorkflowResponse,
+    status_code=status.HTTP_200_OK,
+    summary="Apply lotus-manage Run Workflow Action by Correlation Id",
+    description=(
+        "Applies one workflow action for a discretionary mandate rebalance run resolved by "
+        "submitted correlation id and returns updated workflow state. Use this endpoint when the "
+        "run id is not available but the submitted run correlation id is. This endpoint does not "
+        "accept query parameters."
+    ),
+    responses=_WORKFLOW_ACTION_RESPONSES,
+)
+def apply_dpm_run_workflow_action_by_correlation(
+    request: Request,
+    correlation_id: Annotated[
+        str,
+        Path(
+            description="Correlation identifier used on run submission.",
+            examples=["corr-1234-abcd"],
+        ),
+    ],
+    payload: DpmRunWorkflowActionRequest,
+    service: DpmRunSupportService = shared.Depends(shared.get_dpm_run_support_service),
+    x_correlation_id: Annotated[
+        Optional[str],
+        Header(
+            description="Optional correlation id for workflow action request tracing.",
+            examples=["corr-workflow-action-001"],
+        ),
+    ] = None,
+) -> DpmRunWorkflowResponse:
+    shared._assert_support_apis_enabled()
+    shared._assert_workflow_enabled()
+    shared._reject_unexpected_query_params(request, allowed_params=set())
+    try:
+        response = service.apply_workflow_action_by_correlation(
+            correlation_id=correlation_id,
+            action=payload.action,
+            reason_code=payload.reason_code,
+            comment=payload.comment,
+            actor_id=payload.actor_id,
+            action_correlation_id=x_correlation_id or "c_none",
+        )
+        _record_workflow_action_metric(
+            surface="trace",
+            action=payload.action,
+            outcome="success",
+        )
+        return response
+    except DpmRunNotFoundError as exc:
+        _record_workflow_action_metric(
+            surface="trace",
+            action=payload.action,
+            outcome="not_found",
+        )
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=str(exc)) from exc
+    except DpmWorkflowDisabledError as exc:
+        _record_workflow_action_metric(
+            surface="trace",
+            action=payload.action,
+            outcome="disabled",
+        )
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=str(exc)) from exc
+    except DpmWorkflowTransitionError as exc:
+        _record_workflow_action_metric(
+            surface="trace",
+            action=payload.action,
+            outcome="conflict",
+        )
+        raise HTTPException(status_code=status.HTTP_409_CONFLICT, detail=str(exc)) from exc
+
+
+@shared.router.post(
+    "/rebalance/runs/idempotency/{idempotency_key}/workflow/actions",
+    response_model=DpmRunWorkflowResponse,
+    status_code=status.HTTP_200_OK,
+    summary="Apply lotus-manage Run Workflow Action by Idempotency Key",
+    description=(
+        "Applies one workflow action for a discretionary mandate rebalance run resolved by current "
+        "idempotency-key mapping and returns updated workflow state. Use this endpoint when a "
+        "retry token is the available operational handle. This endpoint does not accept query "
+        "parameters."
+    ),
+    responses=_WORKFLOW_ACTION_RESPONSES,
+)
+def apply_dpm_run_workflow_action_by_idempotency(
+    request: Request,
+    idempotency_key: Annotated[
+        str,
+        Path(
+            description="Idempotency key supplied to `/rebalance/simulate`.",
+            examples=["demo-idem-001"],
+        ),
+    ],
+    payload: DpmRunWorkflowActionRequest,
+    service: DpmRunSupportService = shared.Depends(shared.get_dpm_run_support_service),
+    x_correlation_id: Annotated[
+        Optional[str],
+        Header(
+            description="Optional correlation id for workflow action request tracing.",
+            examples=["corr-workflow-action-002"],
+        ),
+    ] = None,
+) -> DpmRunWorkflowResponse:
+    shared._assert_support_apis_enabled()
+    shared._assert_workflow_enabled()
+    shared._reject_unexpected_query_params(request, allowed_params=set())
+    try:
+        response = service.apply_workflow_action_by_idempotency(
+            idempotency_key=idempotency_key,
+            action=payload.action,
+            reason_code=payload.reason_code,
+            comment=payload.comment,
+            actor_id=payload.actor_id,
+            action_correlation_id=x_correlation_id or "c_none",
+        )
+        _record_workflow_action_metric(
+            surface="retry",
+            action=payload.action,
+            outcome="success",
+        )
+        return response
+    except DpmRunNotFoundError as exc:
+        _record_workflow_action_metric(
+            surface="retry",
+            action=payload.action,
+            outcome="not_found",
+        )
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=str(exc)) from exc
+    except DpmWorkflowDisabledError as exc:
+        _record_workflow_action_metric(
+            surface="retry",
+            action=payload.action,
+            outcome="disabled",
+        )
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=str(exc)) from exc
+    except DpmWorkflowTransitionError as exc:
+        _record_workflow_action_metric(
+            surface="retry",
+            action=payload.action,
+            outcome="conflict",
+        )
+        raise HTTPException(status_code=status.HTTP_409_CONFLICT, detail=str(exc)) from exc
